+++ jest.config.mjs
@@ -182,8 +182,9 @@ const config = {
     // An array of regexp pattern strings that are matched against all modules before the module loader will automatically return a mock for them
     // unmockedModulePathPatterns: undefined,
 
-    // Indicates whether each individual test should be reported during the run
-    // verbose: undefined,
+    // Указывает, следует ли сообщать о каждом отдельном тесте во время выполнения.
+    // Все ошибки также будут отображаться внизу после выполнения.
+    verbose: true,
 
     // An array of regexp patterns that are matched against all source file paths before re-running tests in watch mode
     // watchPathIgnorePatterns: [],