+++ src/contracts.mjs
@@ -1,3 +1,5 @@
 export { default as Interface } from './contracts/Interface.mjs';
 export { default as Traversable } from './contracts/Traversable.mjs';
 export { default as ArrayAccess } from './contracts/ArrayAccess.mjs';
+export { default as Stringable } from './contracts/Stringable.mjs';
+export { default as Throwable } from './contracts/Throwable.mjs';