+++ src/exceptions.mjs
@@ -1,0 +1,1 @@
+export { default as PHPJSError } from './exceptions/PHPJSError.mjs';