+++ tests/contracts.test.mjs
@@ -1,8 +1,10 @@
 import { expect, test } from '@jest/globals';
-import { ArrayAccess, Interface, Traversable } from '../src/contracts.mjs';
+import { ArrayAccess, Interface, Stringable, Throwable, Traversable } from '../src/contracts.mjs';
 
 test('contracts', () => {
     expect(Interface).not.toBeUndefined();
     expect(Traversable).not.toBeUndefined();
     expect(ArrayAccess).not.toBeUndefined();
+    expect(Stringable).not.toBeUndefined();
+    expect(Throwable).not.toBeUndefined();
 });