+++ tests/exceptions.global.test.mjs
@@ -1,0 +1,6 @@
+import { expect, test } from '@jest/globals';
+import '../src/exceptions.global.js';
+
+test('exceptions.global', () => {
+    expect(PHPJSError).not.toBeUndefined();
+});